Extra paremeters for hooks & changed tests & small refactor related to name of one from operations #43
