+++ src/plugins/filters/component/operators.js
@@ -5,7 +5,7 @@ import BaseComponent from './_base';
 import {getOperationName} from '../logicalOperationRegisterer';
 import {OPERATION_ID as OPERATION_AND} from '../logicalOperations/conjunction';
 import {OPERATION_ID as OPERATION_OR} from '../logicalOperations/disjunction';
-import {OPERATION_ID as OPERATION_OR_THEN_VARIABLE} from '../logicalOperations/disjunctionAndVariable';
+import {OPERATION_ID as OPERATION_OR_THEN_VARIABLE} from '../logicalOperations/disjunctionWithExtraCondition';
 import RadioInputUI from './../ui/radioInput';
 
 const SELECTED_AT_START_ELEMENT_INDEX = 0;
